PRSD-458: fixing tests
