+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/LandlordRegistrationSinglePageTests.kt
@@ -11,25 +11,73 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
+import org.mockito.kotlin.any
 import org.mockito.kotlin.whenever
 import uk.gov.communities.prsdb.webapp.helpers.DateTimeHelper
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.ErrorPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordDashboardPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage.Companion.assertPageIs
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.CheckAnswersPageLandlordRegistration
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.ConfirmIdentityFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.CountryOfResidenceFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.EmailFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.LookupAddressFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.LookupContactAddressFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.ManualAddressFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.ManualContactAddressFormPageLandlordRegistration
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.NameFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.NoAddressFoundFormPageLandlordRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordRegistrationJourneyPages.NoContactAddressFoundFormPageLandlordRegistration
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.VerifiedIdentityModel
 import uk.gov.communities.prsdb.webapp.testHelpers.extensions.getFormattedInternationalPhoneNumber
+import java.time.LocalDate
 
 class LandlordRegistrationSinglePageTests : SinglePageTestWithSeedData("data-mockuser-not-landlord.sql") {
     private val phoneNumberUtil = PhoneNumberUtil.getInstance()
+
+    @Nested
+    inner class LandlordRegistrationStartPage {
+        @Test
+        fun `registerAsALandlord page renders`(page: Page) {
+            val landlordRegistrationStartPage = navigator.goToLandlordRegistrationStartPage()
+            BaseComponent.assertThat(landlordRegistrationStartPage.heading).containsText("Private Rented Sector Database")
+        }
+
+        @Test
+        fun `the 'Start Now' button directs an unverified user to the landlord registration name page`(page: Page) {
+            whenever(identityService.getVerifiedIdentityData(any())).thenReturn(null)
+            val landlordRegistrationStartPage = navigator.goToLandlordRegistrationStartPage()
+            landlordRegistrationStartPage.startButton.clickAndWait()
+            assertPageIs(page, NameFormPageLandlordRegistration::class)
+        }
+
+        @Test
+        fun `the 'Start Now' button directs a verified user to the identity confirmation page`(page: Page) {
+            val verifiedIdentityMap =
+                mutableMapOf<String, Any?>(
+                    VerifiedIdentityModel.NAME_KEY to "name",
+                    VerifiedIdentityModel.BIRTH_DATE_KEY to LocalDate.now(),
+                )
+            whenever(identityService.getVerifiedIdentityData(any())).thenReturn(verifiedIdentityMap)
+
+            val landlordRegistrationStartPage = navigator.goToLandlordRegistrationStartPage()
+            landlordRegistrationStartPage.startButton.clickAndWait()
+            assertPageIs(page, ConfirmIdentityFormPageLandlordRegistration::class)
+        }
+    }
+
+    @Nested
+    inner class AlreadyRegistered : NestedSinglePageTestWithSeedData("data-local.sql") {
+        @Test
+        fun `the 'Start Now' button directs a registered landlord to the landlord dashboard page`(page: Page) {
+            val startPage = navigator.goToLandlordRegistrationStartPage()
+            startPage.startButton.clickAndWait()
+            val dashboardPage = assertPageIs(page, LandlordDashboardPage::class)
+            BaseComponent.assertThat(dashboardPage.dashboardBannerHeading).containsText("Alexander Smith")
+        }
+    }
 
     @Nested
     inner class LandlordRegistrationStepVerifyIdentity : NestedSinglePageTestWithSeedData("data-local.sql") {
